Finally using node
-Added node/express code to mad_server.js and got it to run a clientless game!
-Currently spits out the game state into the console each turn. The game continued until the bot beat me.
-In order for a user to watch the game, I believe we next need to set up the socket, then connect the client to it, then get the client to process the incoming game state data json string dump. After that, connect the code that adds/removes queued moves. And new game requests..
