+++ MadJS/mad_constants.js
@@ -1,37 +1,39 @@
 
-///////////
-// Shared constants
-///////////
+// ///////////
+// // Shared constants
+// ///////////
 
-const TERRAIN_TYPE_WATER = 101;
-const TERRAIN_TYPE_SWAMP = 104 ;
-const TERRAIN_TYPE_MOUNTAIN = 105;
-const TERRAIN_TYPE_MOUNTAIN_CRACKED = 106;
-const TERRAIN_TYPE_MOUNTAIN_BROKEN = 107;
+// const TERRAIN_TYPE_WATER = 101;
+// const TERRAIN_TYPE_SWAMP = 104 ;
+// const TERRAIN_TYPE_MOUNTAIN = 105;
+// const TERRAIN_TYPE_MOUNTAIN_CRACKED = 106;
+// const TERRAIN_TYPE_MOUNTAIN_BROKEN = 107;
 
-const ENTITY_TYPE_ADMIRAL = 200;
-const ENTITY_TYPE_SHIP = 201;
-const ENTITY_TYPE_SHIP_2 = 202 // combine 2 ships to make this. Increased growth rate;
-const ENTITY_TYPE_SHIP_3 = 203 // combine 1 ship_2 with a ship to make this. Increased growth rate;
-const ENTITY_TYPE_SHIP_4 = 204 // combine 2 ship_2s or 1 ship_3 and 1 ship to make this. Increased growth rate;
-// const ENTITY_TYPE_INFANTRY = 205;
+// const ENTITY_TYPE_ADMIRAL = 200;
+// const ENTITY_TYPE_SHIP = 201;
+// const ENTITY_TYPE_SHIP_2 = 202 // combine 2 ships to make this. Increased growth rate;
+// const ENTITY_TYPE_SHIP_3 = 203 // combine 1 ship_2 with a ship to make this. Increased growth rate;
+// const ENTITY_TYPE_SHIP_4 = 204 // combine 2 ship_2s or 1 ship_3 and 1 ship to make this. Increased growth rate;
+// // const ENTITY_TYPE_INFANTRY = 205;
 
-const ACTION_MOVE_NORMAL = 1;
-const ACTION_MOVE_HALF = 2;
-const ACTION_MOVE_ALL = 3;
-const ACTION_MOVE_CITY = 4;
-const ACTION_MOVE_NONE = 5;
+// const ACTION_MOVE_NORMAL = 1;
+// const ACTION_MOVE_HALF = 2;
+// const ACTION_MOVE_ALL = 3;
+// const ACTION_MOVE_CITY = 4;
+// const ACTION_MOVE_NONE = 5;
 
 
-const GAME_MODE_FFA = 1;
-const GAME_MODE_FFA_CUST = 2;
-const GAME_MODE_REPLAY = 3;
+// const GAME_MODE_FFA = 1;
+// const GAME_MODE_FFA_CUST = 2;
+// const GAME_MODE_REPLAY = 3;
 
-const GAME_STATUS_INIT = 0; // loading;
-const GAME_STATUS_READY = 1; // able to start;
-const GAME_STATUS_IN_PROGRESS = 2; //;
-const GAME_STATUS_PAUSE = 3; //;
-const GAME_STATUS_GAME_OVER_WIN = 4; // game instance is complete and can no longer be played;
-const GAME_STATUS_GAME_OVER_LOSE = 5; // game instance is complete and can no longer be played;
+// const GAME_STATUS_INIT = 0; // loading;
+// const GAME_STATUS_READY = 1; // able to start;
+// const GAME_STATUS_IN_PROGRESS = 2; //;
+// const GAME_STATUS_PAUSE = 3; //;
+// const GAME_STATUS_GAME_OVER_WIN = 4; // game instance is complete and can no longer be played;
+// const GAME_STATUS_GAME_OVER_LOSE = 5; // game instance is complete and can no longer be played;
 
-const MIN_DISTANCE_ADMIRALS = 5;
+// const MIN_DISTANCE_ADMIRALS = 5;
+
+// const DEFAULT_TICK_SPEED = 500;
